requirements.txt
streamlit==1.28.0
langgraph==0.0.26
langchain==0.1.0
langchain-groq==0.1.0
langchain-community==0.0.13
chromadb==0.4.22
sentence-transformers==2.2.2
reportlab==4.0.8
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
plotly==5.17.0
Pillow==10.1.0
